Tutorial on cross-platform development created for Borland Convention, 2002. A bit dated now.

+++ doors/ny2008/src/nyibbs.cpp
@@ -608,7 +608,7 @@ void            RefreshBest(void) {
 
 
 
-int             main(int argc, char *argv[]) {
+INT16             main(INT16 argc, char *argv[]) {
 	FILE           *justfile, *njustfile;
 	//char          key;
 	//WORD  uintval;
